+++ tests/js/babel.config.js
@@ -1,5 +1,3 @@
-<?php
-
 /**
  * ---------------------------------------------------------------------
  *
@@ -33,15 +31,9 @@
  * ---------------------------------------------------------------------
  */
 
-$AJAX_INCLUDE = 1;
-include('../inc/includes.php');
-header("Content-Type: text/html; charset=UTF-8");
-Html::header_nocache();
-
-Session::checkLoginUser();
-
-try {
-    echo json_encode(Html::getMenuFuzzySearchList(), JSON_THROW_ON_ERROR);
-} catch (JsonException $e) {
-    die(500);
-}
+module.exports = {
+    presets: ['@babel/preset-env'],
+    "plugins": [
+        "@babel/plugin-transform-modules-commonjs"
+    ]
+};